Adjust theme a little, add alias nve
nve stands for nvim edit and utilized the nvd alias to open my vim
config in a way that I can see changes prior to running nix. The nvd
alias is short for nvim dots and is just a shortcut to opening the
configs within my  dots repo and having that nvim session use those
configs directly instead of the configs in the default location.