Setup a Pi 3, fix hetznix02
This focuses on being able to build a SD card image for a Pi 3 named
raspberry and the cofig it contains. As part of making that happen, I
found that I had totally messed up hetznix02's hardware config. I fixed
this as it is where I am builiding the SD image. This work also resulted
in adding a "volume" to hetznix02 and moving `/nix` onto it.