Switching to a built-in theme for a little bit
This is to see if some issues I have been seeing in my terminal are
related to oh-my-posh or to a bug in my custom theme. As a bonus, it is
a good excuse to checkout ways I might want to make my theme better, or
give me ideas for an alternate custom theme.